Updated Ventoy2DiskUpdate.sh

+++ shl/Ventoy2DiskUpdate.sh
@@ -21,19 +21,22 @@ case $architecture in
 	*) echo "=> Unsupported architecture.";exit;;
 esac
 
-if [ -n "$Ventoy2DiskLatestGitHubReleaseURL" ];then
-	echo "=> Ventoy2DiskLatestGitHubReleaseURL = <$Ventoy2DiskLatestGitHubReleaseURL>"
-	\curl -Ls "$Ventoy2DiskLatestGitHubReleaseURL" | tar -C /tmp -xz || exit
-	rsync="$(which rsync) -uth -P -z --skip-compress=$RSYNC_SKIP_COMPRESS_LIST"
-	cp2ext234="$rsync -ogpuv -lSH"
-	test -d $ventoyROOT/ || sudo mkdir -pv $ventoyROOT/
-	time sudo $cp2ext234 -r /tmp/ventoy-$ventoyLatestRelease/* /opt/ventoy
-	sudo chown root -R $ventoyROOT/*
-	rm -fr /tmp/ventoy-$ventoyLatestRelease/
-	test -f $ventoyROOT/tool/x86_64/ash && sudo rm -v $ventoyROOT/tool/x86_64/ash
-
-	if cd $ventoyROOT && [ $# != 0 ];then
-		sudo ./Ventoy2Disk.sh -l "$@"
-		tac $ventoyROOT/log.txt | sed "/^#* Ventoy2Disk/q" | tac
+ventoyCurrentVersion=$(<$ventoyROOT/ventoy/version)
+if [ "$ventoyLatestRelease" != "$ventoyCurrentVersion" ];then
+	if [ -n "$Ventoy2DiskLatestGitHubReleaseURL" ];then
+		echo "=> Ventoy2DiskLatestGitHubReleaseURL = <$Ventoy2DiskLatestGitHubReleaseURL>"
+		\curl -Ls "$Ventoy2DiskLatestGitHubReleaseURL" | tar -C /tmp -xz || exit
+		rsync="$(which rsync) -uth -P -z --skip-compress=$RSYNC_SKIP_COMPRESS_LIST"
+		cp2ext234="$rsync -ogpuv -lSH"
+		test -d $ventoyROOT/ || sudo mkdir -pv $ventoyROOT/
+		time sudo $cp2ext234 -r /tmp/ventoy-$ventoyLatestRelease/* /opt/ventoy
+		sudo chown root -R $ventoyROOT/*
+		rm -fr /tmp/ventoy-$ventoyLatestRelease/
+		test -f $ventoyROOT/tool/x86_64/ash && sudo rm -v $ventoyROOT/tool/x86_64/ash
+	
+		if cd $ventoyROOT && [ $# != 0 ];then
+			sudo ./Ventoy2Disk.sh -l "$@"
+			tac $ventoyROOT/log.txt | sed "/^#* Ventoy2Disk/q" | tac
+		fi
 	fi
 fi
